requirements.txt
python-dotenv==0.12.0
requests==2.21.0